requirements.txt
Flask
qrcode[pil]
python-dotenv
